requirements.txt
streamlit==1.10.0
Pillow==9.1.1
numpy
torch
torchvision
pandas
opencv-python
rapidocr_onnxruntime


#opencv-python-headless<4.3
opencv-python